moved query parsing into own module
